refactor pubSub module
