+++ server/pubSub/index.js
@@ -1,23 +1,9 @@
-const PubSub = require('pubsub-js');
 const events = require('./eventTypes');
-const Sentry = require('../sentry');
+const { emit, listen } = require('./createPubSub');
 
-const emit = (event, data) => {
-  console.log('published', { event, data });
-  PubSub.publish(event, data);
-};
-
-const listen = (event, listener) => {
-  console.log('listen', { event, listener });
-
-  PubSub.subscribe(event, async (_event, data) => {
-    try {
-      await listener(data);
-    } catch (error) {
-      Sentry.captureException(error);
-    }
-  });
-};
+require('./booking');
+require('./profile');
+require('./user');
 
 module.exports = {
   emit,
